Reorganise headers
