+++ R/compile.R
@@ -130,18 +130,20 @@ dust_generate <- function(config, filename, linking_to, cpp_std,
 
 
 dust_generate_cpp <- function(system, config, data) {
-  includes <- c("cpp11.hpp", "dust2/r/discrete.hpp")
-  code <- substitute_dust_template(data, "dust_core.cpp")
+  includes <- c("cpp11.hpp", "dust2/r/discrete/system.hpp")
+  code <- substitute_dust_template(data, "discrete/system.cpp")
 
   header <- sprintf(
     "// Generated by dust2 (version %s) - do not edit",
     data$dust_version)
 
   if (config$has_compare) {
-    includes <- c(includes, "dust2/r/filter.hpp", "dust2/r/unfilter.hpp")
+    includes <- c(includes,
+                  "dust2/r/discrete/filter.hpp",
+                  "dust2/r/discrete/unfilter.hpp")
     code <- c(code,
               "",
-              substitute_dust_template(data, "dust_compare.cpp"))
+              substitute_dust_template(data, "discrete/compare.cpp"))
   }
 
   c(header,
