+++ inst/include/dust2/discrete/system.hpp
