+++ inst/include/dust2/r/discrete/unfilter.hpp
@@ -2,7 +2,7 @@
 
 #include <mcstate/r/random.hpp>
 #include <dust2/r/helpers.hpp>
-#include <dust2/unfilter.hpp>
+#include <dust2/discrete/unfilter.hpp>
 
 namespace dust2 {
 namespace r {
