+++ inst/template/discrete/compare.cpp
