+++ src/sir.cpp
@@ -125,9 +125,9 @@ public:
 };
 
 #include <cpp11.hpp>
-#include <dust2/r/discrete.hpp>
-#include <dust2/r/filter.hpp>
-#include <dust2/r/unfilter.hpp>
+#include <dust2/r/discrete/system.hpp>
+#include <dust2/r/discrete/filter.hpp>
+#include <dust2/r/discrete/unfilter.hpp>
 
 [[cpp11::register]]
 SEXP dust2_discrete_sir_alloc(cpp11::list r_pars, cpp11::sexp r_time, cpp11::sexp r_dt, cpp11::sexp r_n_particles, cpp11::sexp r_n_groups, cpp11::sexp r_seed, cpp11::sexp r_deterministic) {
